WLAN Gen changes for RFsgCfgClockDelay and correcting BTGen custom file

+++ source/codegen/metadata/nirfmxwlangen/attributes.py
@@ -796,7 +796,7 @@ attributes = {
     211: {
         'access': 'read-write',
         'name': 'OVERSAMPLING_RATIO',
-        'type': 'int32'
+        'type': 'float64'
     },
     212: {
         'access': 'read-write',
